Move `std::io::Seek` to `core::io`
