+++ library/alloc/src/io/mod.rs
@@ -13,9 +13,9 @@ pub use core::io::const_error;
 pub use core::io::{BorrowedBuf, BorrowedCursor};
 #[unstable(feature = "alloc_io", issue = "154046")]
 pub use core::io::{
-    Chain, Cursor, Empty, Error, ErrorKind, IoSlice, IoSliceMut, Repeat, Result, SeekFrom, Sink,
-    Take, empty, repeat, sink,
+    Chain, Cursor, Empty, Error, ErrorKind, IoSlice, IoSliceMut, Repeat, Result, Seek, SeekFrom,
+    Sink, Take, empty, repeat, sink,
 };
 #[doc(hidden)]
 #[unstable(feature = "core_io_internals", reason = "exposed only for libstd", issue = "none")]
-pub use core::io::{IoHandle, OsFunctions, SizeHint, chain, take};
+pub use core::io::{IoHandle, OsFunctions, SizeHint, chain, stream_len_default, take};
