+++ library/alloc/src/lib.rs
@@ -155,6 +155,7 @@
 #![feature(ptr_metadata)]
 #![feature(raw_os_error_ty)]
 #![feature(rev_into_inner)]
+#![feature(seek_stream_len)]
 #![feature(set_ptr_value)]
 #![feature(share_trait)]
 #![feature(sized_type_properties)]
