+++ library/core/src/io/impls.rs
@@ -1,4 +1,4 @@
-use crate::io::SizeHint;
+use crate::io::{self, Seek, SeekFrom, SizeHint};
 
 // =============================================================================
 // Forwarding implementations
@@ -14,6 +14,34 @@ impl<T> SizeHint for &mut T {
     #[inline]
     fn upper_bound(&self) -> Option<usize> {
         SizeHint::upper_bound(*self)
+    }
+}
+
+#[stable(feature = "rust1", since = "1.0.0")]
+impl<S: Seek + ?Sized> Seek for &mut S {
+    #[inline]
+    fn seek(&mut self, pos: SeekFrom) -> io::Result<u64> {
+        (**self).seek(pos)
+    }
+
+    #[inline]
+    fn rewind(&mut self) -> io::Result<()> {
+        (**self).rewind()
+    }
+
+    #[inline]
+    fn stream_len(&mut self) -> io::Result<u64> {
+        (**self).stream_len()
+    }
+
+    #[inline]
+    fn stream_position(&mut self) -> io::Result<u64> {
+        (**self).stream_position()
+    }
+
+    #[inline]
+    fn seek_relative(&mut self, offset: i64) -> io::Result<()> {
+        (**self).seek_relative(offset)
     }
 }
