+++ library/std/src/io/cursor.rs
@@ -7,42 +7,7 @@ pub use core::io::Cursor;
 use crate::alloc::Allocator;
 use crate::cmp;
 use crate::io::prelude::*;
-use crate::io::{self, BorrowedCursor, ErrorKind, IoSlice, IoSliceMut, SeekFrom};
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<T> io::Seek for Cursor<T>
-where
-    T: AsRef<[u8]>,
-{
-    fn seek(&mut self, style: SeekFrom) -> io::Result<u64> {
-        let (base_pos, offset) = match style {
-            SeekFrom::Start(n) => {
-                self.set_position(n);
-                return Ok(n);
-            }
-            SeekFrom::End(n) => (self.get_ref().as_ref().len() as u64, n),
-            SeekFrom::Current(n) => (self.position(), n),
-        };
-        match base_pos.checked_add_signed(offset) {
-            Some(n) => {
-                self.set_position(n);
-                Ok(n)
-            }
-            None => Err(io::const_error!(
-                ErrorKind::InvalidInput,
-                "invalid seek to a negative or overflowing position",
-            )),
-        }
-    }
-
-    fn stream_len(&mut self) -> io::Result<u64> {
-        Ok(self.get_ref().as_ref().len() as u64)
-    }
-
-    fn stream_position(&mut self) -> io::Result<u64> {
-        Ok(self.position())
-    }
-}
+use crate::io::{self, BorrowedCursor, ErrorKind, IoSlice, IoSliceMut};
 
 #[stable(feature = "rust1", since = "1.0.0")]
 impl<T> Read for Cursor<T>
