+++ library/std/src/io/mod.rs
@@ -299,7 +299,6 @@ mod tests;
 
 use core::slice::memchr;
 
-pub(crate) use alloc_crate::io::IoHandle;
 #[unstable(feature = "raw_os_error_ty", issue = "107792")]
 pub use alloc_crate::io::RawOsError;
 #[doc(hidden)]
@@ -311,8 +310,9 @@ pub use alloc_crate::io::const_error;
 pub use alloc_crate::io::{BorrowedBuf, BorrowedCursor};
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use alloc_crate::io::{
-    Chain, Empty, Error, ErrorKind, Repeat, Result, SeekFrom, Sink, Take, empty, repeat, sink,
+    Chain, Empty, Error, ErrorKind, Repeat, Result, Seek, SeekFrom, Sink, Take, empty, repeat, sink,
 };
+pub(crate) use alloc_crate::io::{IoHandle, stream_len_default};
 #[stable(feature = "iovec", since = "1.36.0")]
 pub use alloc_crate::io::{IoSlice, IoSliceMut};
 use alloc_crate::io::{OsFunctions, SizeHint};
@@ -1762,197 +1762,6 @@ pub trait Write {
     }
 }
 
-/// The `Seek` trait provides a cursor which can be moved within a stream of
-/// bytes.
-///
-/// The stream typically has a fixed size, allowing seeking relative to either
-/// end or the current offset.
-///
-/// # Examples
-///
-/// [`File`]s implement `Seek`:
-///
-/// [`File`]: crate::fs::File
-///
-/// ```no_run
-/// use std::io;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-/// use std::io::SeekFrom;
-///
-/// fn main() -> io::Result<()> {
-///     let mut f = File::open("foo.txt")?;
-///
-///     // move the cursor 42 bytes from the start of the file
-///     f.seek(SeekFrom::Start(42))?;
-///     Ok(())
-/// }
-/// ```
-#[stable(feature = "rust1", since = "1.0.0")]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoSeek")]
-pub trait Seek {
-    /// Seek to an offset, in bytes, in a stream.
-    ///
-    /// A seek beyond the end of a stream is allowed, but behavior is defined
-    /// by the implementation.
-    ///
-    /// If the seek operation completed successfully,
-    /// this method returns the new position from the start of the stream.
-    /// That position can be used later with [`SeekFrom::Start`].
-    ///
-    /// # Errors
-    ///
-    /// Seeking can fail, for example because it might involve flushing a buffer.
-    ///
-    /// Seeking to a negative offset is considered an error.
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn seek(&mut self, pos: SeekFrom) -> Result<u64>;
-
-    /// Rewind to the beginning of a stream.
-    ///
-    /// This is a convenience method, equivalent to `seek(SeekFrom::Start(0))`.
-    ///
-    /// # Errors
-    ///
-    /// Rewinding can fail, for example because it might involve flushing a buffer.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::io::{Read, Seek, Write};
-    /// use std::fs::OpenOptions;
-    ///
-    /// let mut f = OpenOptions::new()
-    ///     .write(true)
-    ///     .read(true)
-    ///     .create(true)
-    ///     .open("foo.txt")?;
-    ///
-    /// let hello = "Hello!\n";
-    /// write!(f, "{hello}")?;
-    /// f.rewind()?;
-    ///
-    /// let mut buf = String::new();
-    /// f.read_to_string(&mut buf)?;
-    /// assert_eq!(&buf, hello);
-    /// # std::io::Result::Ok(())
-    /// ```
-    #[stable(feature = "seek_rewind", since = "1.55.0")]
-    fn rewind(&mut self) -> Result<()> {
-        self.seek(SeekFrom::Start(0))?;
-        Ok(())
-    }
-
-    /// Returns the length of this stream (in bytes).
-    ///
-    /// The default implementation uses up to three seek operations. If this
-    /// method returns successfully, the seek position is unchanged (i.e. the
-    /// position before calling this method is the same as afterwards).
-    /// However, if this method returns an error, the seek position is
-    /// unspecified.
-    ///
-    /// If you need to obtain the length of *many* streams and you don't care
-    /// about the seek position afterwards, you can reduce the number of seek
-    /// operations by simply calling `seek(SeekFrom::End(0))` and using its
-    /// return value (it is also the stream length).
-    ///
-    /// Note that length of a stream can change over time (for example, when
-    /// data is appended to a file). So calling this method multiple times does
-    /// not necessarily return the same length each time.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// #![feature(seek_stream_len)]
-    /// use std::{
-    ///     io::{self, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///
-    ///     let len = f.stream_len()?;
-    ///     println!("The file is currently {len} bytes long");
-    ///     Ok(())
-    /// }
-    /// ```
-    #[unstable(feature = "seek_stream_len", issue = "59359")]
-    fn stream_len(&mut self) -> Result<u64> {
-        stream_len_default(self)
-    }
-
-    /// Returns the current seek position from the start of the stream.
-    ///
-    /// This is equivalent to `self.seek(SeekFrom::Current(0))`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::{
-    ///     io::{self, BufRead, BufReader, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = BufReader::new(File::open("foo.txt")?);
-    ///
-    ///     let before = f.stream_position()?;
-    ///     f.read_line(&mut String::new())?;
-    ///     let after = f.stream_position()?;
-    ///
-    ///     println!("The first line was {} bytes long", after - before);
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "seek_convenience", since = "1.51.0")]
-    fn stream_position(&mut self) -> Result<u64> {
-        self.seek(SeekFrom::Current(0))
-    }
-
-    /// Seeks relative to the current position.
-    ///
-    /// This is equivalent to `self.seek(SeekFrom::Current(offset))` but
-    /// doesn't return the new position which can allow some implementations
-    /// such as [`BufReader`] to perform more efficient seeks.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::{
-    ///     io::{self, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     f.seek_relative(10)?;
-    ///     assert_eq!(f.stream_position()?, 10);
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// [`BufReader`]: crate::io::BufReader
-    #[stable(feature = "seek_seek_relative", since = "1.80.0")]
-    fn seek_relative(&mut self, offset: i64) -> Result<()> {
-        self.seek(SeekFrom::Current(offset))?;
-        Ok(())
-    }
-}
-
-pub(crate) fn stream_len_default<T: Seek + ?Sized>(self_: &mut T) -> Result<u64> {
-    let old_pos = self_.stream_position()?;
-    let len = self_.seek(SeekFrom::End(0))?;
-
-    // Avoid seeking a third time when we were already at the end of the
-    // stream. The branch is usually way cheaper than a seek operation.
-    if old_pos != len {
-        self_.seek(SeekFrom::Start(old_pos))?;
-    }
-
-    Ok(len)
-}
-
 fn read_until<R: BufRead + ?Sized>(r: &mut R, delim: u8, buf: &mut Vec<u8>) -> Result<usize> {
     let mut read = 0;
     loop {
@@ -2601,49 +2410,6 @@ impl<T: BufRead> BufRead for Take<T> {
         let amt = cmp::min(amt as u64, self.limit) as usize;
         self.limit -= amt as u64;
         self.inner.consume(amt);
-    }
-}
-
-#[stable(feature = "seek_io_take", since = "1.89.0")]
-impl<T: Seek> Seek for Take<T> {
-    fn seek(&mut self, pos: SeekFrom) -> Result<u64> {
-        let new_position = match pos {
-            SeekFrom::Start(v) => Some(v),
-            SeekFrom::Current(v) => self.position().checked_add_signed(v),
-            SeekFrom::End(v) => self.len.checked_add_signed(v),
-        };
-        let new_position = match new_position {
-            Some(v) if v <= self.len => v,
-            _ => return Err(ErrorKind::InvalidInput.into()),
-        };
-        while new_position != self.position() {
-            if let Some(offset) = new_position.checked_signed_diff(self.position()) {
-                self.inner.seek_relative(offset)?;
-                self.limit = self.limit.wrapping_sub(offset as u64);
-                break;
-            }
-            let offset = if new_position > self.position() { i64::MAX } else { i64::MIN };
-            self.inner.seek_relative(offset)?;
-            self.limit = self.limit.wrapping_sub(offset as u64);
-        }
-        Ok(new_position)
-    }
-
-    fn stream_len(&mut self) -> Result<u64> {
-        Ok(self.len)
-    }
-
-    fn stream_position(&mut self) -> Result<u64> {
-        Ok(self.position())
-    }
-
-    fn seek_relative(&mut self, offset: i64) -> Result<()> {
-        if !self.position().checked_add_signed(offset).is_some_and(|p| p <= self.len) {
-            return Err(ErrorKind::InvalidInput.into());
-        }
-        self.inner.seek_relative(offset)?;
-        self.limit = self.limit.wrapping_sub(offset as u64);
-        Ok(())
     }
 }
